Add additional wait condition before clicking complain button
Before there was just a check to see if the loader disappeared before
clicking the complain button. But sometimes the loader disappears for a
short time after login and then reappears again before page is loaded.

Now there is an additional check to see if the dashboard title is
visible, then there is a wait for the loader to disappear. This way,
if the loader reappears after login, it is caught by the wait
condition.